Fix build [DO NOT MERGE]
The branch is missing some CLs for 64-bit support. So, until they are
added, this change will fix the build.

+++ tools/layoutlib/bridge/src/android/content/res/AssetManager_Delegate.java
@@ -29,7 +29,7 @@ public class AssetManager_Delegate {
 
     @LayoutlibDelegate
     /*package*/ static int newTheme(AssetManager manager) {
-        return Resources_Theme_Delegate.getDelegateManager()
+        return (int) Resources_Theme_Delegate.getDelegateManager()
                 .addNewDelegate(new Resources_Theme_Delegate());
     }
